+++ engine/src/main/java/dev/gamekit/core/Window.java
@@ -1,18 +1,27 @@
 package dev.gamekit.core;
 
+import org.apache.logging.log4j.LogManager;
+import org.apache.logging.log4j.Logger;
+
 import javax.swing.*;
 import java.awt.*;
 import java.awt.image.BufferedImage;
 
 /** Singleton class which manages the {@link JFrame} the application is rendered in */
 public final class Window {
+  private static final Logger LOGGER = LogManager.getLogger();
+
+  private static boolean isFullScreen = false;
   private static Resolution resolution = Resolution._800_600;
   private static Window instance;
 
   private final int width;
   private final int height;
+  //  private final Dimension renderSize;
+  //  private final Dimension frameSize;
   private final int centerX;
   private final int centerY;
+  //  private final Point center;
   private final double scaleRatio;
   private final JFrame frame;
   private final JPanel renderPanel;
@@ -23,29 +32,37 @@ public final class Window {
 
   Window(String title) {
     Dimension screenSize = Toolkit.getDefaultToolkit().getScreenSize();
-    width = resolution == Resolution.NATIVE ? screenSize.width : resolution.minWidth;
-    height = resolution == Resolution.NATIVE ? screenSize.height : resolution.minHeight;
+    width = resolution == Resolution.NATIVE ? screenSize.width : resolution.renderWidth;
+    height = resolution == Resolution.NATIVE ? screenSize.height : resolution.renderHeight;
     scaleRatio = Math.min(screenSize.getWidth() / width, screenSize.getHeight() / height);
     centerX = width / 2;
     centerY = height / 2;
 
+    LOGGER.debug("Resolution: {}x{}", width, height);
+    LOGGER.debug("Fullscreen: {}", isFullScreen);
     createRenderTargets();
 
     renderPanel = new JPanel();
     renderPanel.setBackground(Color.BLACK);
 
     frame = new JFrame(title);
+    frame.setMinimumSize(new Dimension(width, height));
     frame.setLocationRelativeTo(null);
-    frame.setUndecorated(true);
     frame.setTitle(title);
+
+    if (Window.isFullScreen) {
+      frame.setUndecorated(true);
+      GraphicsEnvironment
+        .getLocalGraphicsEnvironment()
+        .getDefaultScreenDevice()
+        .setFullScreenWindow(frame);
+    } else {
+      frame.setPreferredSize(new Dimension(width, height));
+      frame.setResizable(false);
+    }
+
     frame.add(renderPanel);
     frame.pack();
-
-    // Enter true full screen mode
-    GraphicsEnvironment
-      .getLocalGraphicsEnvironment()
-      .getDefaultScreenDevice()
-      .setFullScreenWindow(frame);
 
     Window.instance = this;
   }
@@ -65,6 +82,18 @@ public final class Window {
   public static void setResolution(Resolution resolution) {
     if (Window.instance == null) {
       Window.resolution = resolution;
+    }
+  }
+
+  /**
+   * Sets whether the {@link Window} should launch in full screen or in windowed mode
+   * <p>
+   * After the {@code Window} instance is created, this method does nothing
+   * @param fullscreen The fullscreen state
+   */
+  public static void setFullscreen(boolean fullscreen) {
+    if (Window.instance == null) {
+      Window.isFullScreen = fullscreen;
     }
   }
 
@@ -99,16 +128,22 @@ public final class Window {
   Graphics2D getNodeGraphics() { return nodeGraphics; }
 
   void redraw() {
-    int scaledWidth = (int) (width * scaleRatio);
-    int scaledHeight = (int) (height * scaleRatio);
-    int dx1 = (int) (0.5 * (frame.getWidth() - scaledWidth));
-    int dy1 = (int) (0.5 * (frame.getHeight() - scaledHeight));
-    int dx2 = dx1 + scaledWidth;
-    int dy2 = dy1 + scaledHeight;
     Graphics2D renderGraphics = (Graphics2D) renderPanel.getGraphics();
 
-    renderGraphics.drawImage(sceneTarget, dx1, dy1, dx2, dy2, 0, 0, width, height, null);
-    renderGraphics.drawImage(nodeTarget, dx1, dy1, dx2, dx2, 0, 0, width, height, null);
+    if (Window.isFullScreen) {
+      int scaledWidth = (int) (width * scaleRatio);
+      int scaledHeight = (int) (height * scaleRatio);
+      int dx1 = (int) (0.5 * (frame.getWidth() - scaledWidth));
+      int dy1 = (int) (0.5 * (frame.getHeight() - scaledHeight));
+      int dx2 = dx1 + scaledWidth;
+      int dy2 = dy1 + scaledHeight;
+
+      renderGraphics.drawImage(sceneTarget, dx1, dy1, dx2, dy2, 0, 0, width, height, null);
+      renderGraphics.drawImage(nodeTarget, dx1, dy1, dx2, dx2, 0, 0, width, height, null);
+    } else {
+      renderGraphics.drawImage(sceneTarget, null, 0, 0);
+      renderGraphics.drawImage(nodeTarget, null, 0, 0);
+    }
   }
 
   void createRenderTargets() {
@@ -132,14 +167,14 @@ public final class Window {
     public static final Resolution _1920_1080 = new Resolution(1920, 1080);
     public static final Resolution NATIVE = new Resolution(-1, -1);
 
-    public final int minWidth;
-    public final int minHeight;
+    public final int renderWidth;
+    public final int renderHeight;
     public final double ratio;
 
-    public Resolution(int minWidth, int minHeight) {
-      this.minWidth = minWidth;
-      this.minHeight = minHeight;
-      ratio = (double) minWidth / minHeight;
+    public Resolution(int renderWidth, int renderHeight) {
+      this.renderWidth = renderWidth;
+      this.renderHeight = renderHeight;
+      ratio = (double) renderWidth / renderHeight;
     }
   }
 }